lint: fix linting
